[fvm] Define board/product level dimension args
Allows callers of generate_fvm() to specify blobfs/minfs dimensions on a
product level or a board level.

If product-level arguments are passed, they are preferred over
board-level arguments. This is sufficient for our current use cases, but
in the future we may want to use board-level as a minimum setting since
the board would presumably determine the physical limits on the
partition dimensions (which the product could further restrict if
desired).

This change is a NOP for the current build, but allows a soft transition
onto these new board/product dimension arguments. After all clients are
onto the new arguments, we can clean up the old arguments.

Eventually we will remove the different configuration for sparse blobfs
instances, and use a product-level override to accomplish the same goal.

Bug: 67015
Change-Id: I9478c6dce6b537e8a574c96046fd0b9686ca14ec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465989
Reviewed-by: Vikram Auradkar <auradkar@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>