[kernel][debuglog] Avoid shutdown race
It is possible, albeit unlikely, for multiple threads to attempt to
shutdown the debuglog. One way this could occur could be for the system
to OOM and at the same time the critical job to be killed.

The solution here is to detect multiple shutdown requests, and have the
second request simply wait its maximum possible time.

Bug: 57008

Change-Id: I9e289672af39ec466f6e11e7aef9b4269c44040e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468154
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>