[display][hdmi][i915] Add audio support
We pass ELD (EDID like data as defined in the Intel HDA spec) from
the display controller (gets it from the EDID) to the display
implementationdriver (i915). For this we:

1. Create a small ELD library to be used by HDMI display drivers to
be shared with any future HDA audio HDMI drivers.
2. Add a SetEld banjo method independent from ApplyConfiguration only
to be called when the ELDs change.
3. Extract ELDs from EDIDs (right after EDID reads) and call SetEld
so display implementation drivers can make use of them (we use it
in the i915).
4. Update the fake display driver to provide canned EDIDs via I2C.
5. Trigger HDA ELD functionality in the i915 driver to be used by an
audio HDMI driver (in an upcoming CL).

Test: display-core-unittests
Test: Manual HDMI audio check with upcoming HDA audio CL.
Change-Id: I75e7d1dd91ba47f6035e650d49ae387e5ed16308
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467857
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Reviewed-by: John Bauman <jbauman@google.com>