[memalloc] Add RangeStorage class for allocating storage space.
The current memalloc::Allocator API requires a fbl::Span<Range> of
pre-allocated storage to be passed in to the constructor. This storage
is used as the internal range nodes of the allocator.

However, the Range type is not POD, meaning that global/static
allocations need to call the constructor when we first use the storage,
and emit calls to the destructor on program termination:

  static memalloc::Range storage[100];  // non-POD
  memalloc::Allocator allocator(storage);

This isn't ideal, and causes link errors in minimal environments such as
physboot which don't have the appropriate symbols defined to support
global destruction.

Instead, we introduce a new RangeStorage class that can be treated like
a POD class and avoids these issues, and leave explicit
construction/destruction up to the memalloc::Allocator class:

  static memalloc::RangeStorage storage[100];
  memalloc::Allocator allocator(storage);

Change-Id: Ie45a2a7f083560b023e38ee7510dccf80a5e1c68
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465096
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>