[system-update-committer] don't commit if current is already healthy
In this change, we modify the system-update-committer so that if the
current slot is healthy, we don't need to commit. Having it commit when
current is healthy (the previous behavior) isn't functionally any different because the paver calls are just no-ops. However, it makes the code complicated.

To optimize for simplicity, let's change it so that when current is
healthy, we need not commit. Also, we can rename things to make them
more intuitive.

Test: fx test -o system-update-committer-tests
Test: fx test -o system-update-committer-integration-tests
Test: fx test -o system-update-checker-integration-tests
Test: fx test -o omaha-client-integration-tests
Fixed: 67237
Change-Id: I57f48613e53631f79bbcfc0df73aa6d88d2822db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468134
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>