[build] Port 'kasan' build variant to Fuchdsia build.
This CL implements the 'kasan' build variant in the Fuchsia build.
It is used to enable Address Sanitization in the kernel itself. It
is similar to the "asan" variant, with the exception that it
currently doesn't support arm64 builds (see https://fxbug.dev/30033).

This also means that enabling the "asan" variant should not touch
the kernel. To achieve this, new variant tags 'kernel-only' and
'kernel-excluded' are introduced, and variant_toolchain_suite() is
slightly modified to add either one to toolchain_variant.tags
adequately.

In the future, i.e. when 30033 is fixed, "kasan" and "kasan-sancov"
might be merged into "asan" and "asan-sancov" directly, removing
the need for this special case.

- build/toolchain/variant_tags.gni: Add 'kernel-only' and
  'kernel-excluded' and document that they are treated specially by
  variant_toolchain_suite().

- build/config/BUILDCONFIG.gn: Add 'kernel-excluded' tags to
  'asan' and 'ubsan' variants to ensure they are never applied
  to the kernel toolchains. Also add 'kasan' and 'kasan-sancov'
  as new 'kernel-only' build variants. This also removes the
  old shorthands with the same name, which used to ignore these
  values from 'select_variant'.

  Note that before and after the CL, a 'kasan' value in
  select_variant is still passed to the Zircon build, where
  the kernel is still being built by default for now, so this
  doesn't impact the build.

+ Update the 'profile' variant to work correctly with the kernel.

For testing, add //zircon/kernel to dev_build_only_deps in
your args.gn to ensure the Fuchsia build also builds the kernel
image, and build with 'kasan', 'asan', etc to verify that they
are, or are not, affected by the build variant choice.

Bug: 66830
Bug: 30033

Change-Id: I710ad692a9af28885e9c1c3f58939b35dc8af910
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465993
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Commit-Queue: David Turner <digit@google.com>