[kernel][debug] add command to display build instrumentation.
This CL adds a k build_instrumentation command, which prints a list of
the build instrumentation features available in this kernel build.

The list is non-exhaustive as each new instrumentation has to be
manually added.

This command is intended to be used by host tests so they can check if a
given instrumentation is enabled or not before running a test.

The output looks like this:

```
$ k build_instrumentation
[00024.594] 01652:02533> build_instrumentation: address_sanitizer
[00024.594] 01652:02533> build_instrumentation: debug_assert
[00024.594] 01652:02533> build_instrumentation: done
```

BUG=66578

Change-Id: I99142d32a91e3f6b4ac09f4c138bf71da565b648
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464984
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>