[component_manager] Add a Discover action
This adds a Discover action, and dispatches the Discovered event under
this action. This allows us to guarantee that Discovered is always
dispatched before Resolved, and is never double-dispatched.

As a drive by, remove some errors around invalid environments that
aren't actually possible and replace them with unwraps. The errors in
question arise if something is wrong in the environment FIDL, but this
shouldn't be possible the FIDL is already validated (which we assume
elsewhere).

Change-Id: Ic6bc951c2dfa527db9e6f64fd71391ae80b5e2c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463604
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Gary Bressler <geb@google.com>