[src/test/emulator] Skip PMM crash tests on KASAN builds.
This CL modifies the emulator code to allow for checking whether any
message from a given list is present in the kernel log output.

The PMM Crash tests performs an UAF, which in kasan builds is detected
by the sanitizer instrumentation.

The test now checks whether the kernel has been built with address
sanitizer instrumentation before running, and is skipped if so.

To run this test:

```shell
$ fx test kernel_panic_tests
```

BUG=66578

Change-Id: I3c53220a9977fc96afc3790df97423fd4fdc4eca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464985
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>