[kernel][libc] Fix printf bug
We were treating precision as width - and if both width and precision
were specified, the latter would win and the former would not be
applied. A simple manifestation of this bug was the inability to apply
width to string_view printing.

Change-Id: I88e4aec6360247ad30a90ddc7c433a619d175059
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465616
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>