[pm][serve] Wait on goroutines in pm serve
This change explicitly waits on goroutines produced by `serve.Run()`
to exit before returning.

fxbug.dev/66732 noted a case where the pm serve test flaked because the
Go testing package couldn't remove a temporary directory created by the
test.

The particular error was:
```
testing.go:915: TempDir RemoveAll cleanup: unlinkat /.../TestServer713674168/002: directory not empty
```

The `002` here is significant: it indicates that it was the repository
directory that was not empty (Go's testing package uses sequential
numbers for the temporary directories it creates to help identify these
sorts of cases). The test tools themselves only ever call
`ioutil.ReadAll` and an audit of the `go-tuf` package didn't indicate
any leaked file descriptors. This leaves the HTTP server itself as the
likely culprit.

In fxrev.dev/433288, a new goroutine was added to poll metadata events
on a secondly basis. Subsequently, fxrev.dev/460574 removed the polling
by adding another watcher API, and the changed goroutine was not waited
on before function exit.

The hypothesis for this change is that this goroutine was still running
when `serve.Run()` returned to the test, and still when the test
successfully returned to the test runner. In the process of running, and
while the test is cleaning up, one of the `ioutil.WriteFile()` cases
produces a new file to a path that `os.RemoveAll()` had already handled.

Test: pm_cmd_pm_serve_test still passes, though this isn't necessarily
indicative of a fix. Reproducing this race is difficult, even with
instrumentation, so this change is optimistic that the flake will stop.

Fixed: 66732
Change-Id: Iab0d0373606107c1bfdd26316ead8aaebb0e4929
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464494
Commit-Queue: Devon H. O'Dell <dhobsd@google.com>
Reviewed-by: Kevin Wells <kevinwells@google.com>