[runtests] fix invalid memory accesses.
The arg parsing logic in runtests had an off-by-one error. Valid
arguments from argv are in the range [0, argc).

When the last argument was a flag with an extra parameter (for
example `-n`), `optind` would be equal to `argc`, however the check
only compares whether `optind` is greater than `argc`.

The resulting logic was that if a flag was missing, most of the checks
would pass regardless, and garbage data would be read.

To reproduce this bug, you can run `runtests -n`, without an extra
parameter.

Change-Id: I416dc4bc23776fa0da59a0c501d9656522405cda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469197
Reviewed-by: Gary Miguel <garymm@google.com>
Commit-Queue: Marco Vanotti <mvanotti@google.com>