[fidl][llcpp] Generated code uses typed channels
This CL updates LLCPP generated code to use typed channel endpoints.

Implicit conversions and cascading inheritance are used to reduce the
extent of the breaking changes. They will be removed as we complete the
migration.

- Existing code using the client API remain unchanged thanks to the
implicit conversion from a channel to a typed endpoint.

- Code moving a channel from an LLCPP domain object into another channel
are affected by this CL, since there is no way to define an implicit
conversion from a ClientEnd/ServerEnd into a zircon channel. I've
applied fixes using a mix of tool-assisted migration and manual changes.
Fortunately the number of these uses is not that large.

- Existing code using the server interface remain mostly unchanged
thanks to the cascading inheritance. There are a few cases where our
compiler emitted a warning (treated as error) when the user declares a
method with the same name as one of the virtual methods. Those are
resolved by inserting a `using Interface::FooMethod` statement to
disambiguate.

- Users of the server unbind handler are unchanged because the unbind
handler still returns a raw channel. Those will be migrated in a future
CL.

Bug: 65212
Change-Id: I5bc01eca91c86b276f034a26f5722f07db59b695
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/464814
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Vincent Belliard <vbelliard@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>