[fshost] Add build dependency for enabling the decompression sandbox
Adds in the option for sandboxing decompression to the fshost config and
plumbs it through to the mount options.

This currently adds a small rule that also includes gpt_all, mainly to
support x64 which currently uses this option as well. This will probably
change anyways when I actually go to enable this for something.

Change-Id: I894724d4d17ff7b8a7f56b42cedf8efde5138a0b
Bug: 62177
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/463821
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Chris Suter <csuter@google.com>