[proc][bionic] Add toolchain and declaration
This CL addins a GN target for building bionic in the Fuchsia source
tree. We do not currently have the source code for bionic in the Fuchsia
tree, but we are working towards adding it.

This eventual goal is to be able to run the bionic test suite against
Fuchisa to validate our *nix compatibility.

Bug: 66359
Change-Id: I5ddffea2822531b53b2340753aada633ed7e5096
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468632
Commit-Queue: Adam Barth <abarth@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: P.Y. Laligand <pylaligand@google.com>