[audio_core] Throttle RPCs to cobalt
See comments on the bug. This CL will send at most 30 RPCs every minute.

Also use the CobaltLogger class which sends cobalt events sequentially
to avoid channel overflow. This class is thread-safe and handles
buffering of pending events. (As noted on the bug, it doesn't actually
bound the buffer size, which means we are potentially suspect to OOM,
but I'll fix that in a follow-up CL.)

Bug: 67416
Change-Id: I7672a8f735f2fce21e27fe8c9f37d19180083e85
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469114
Reviewed-by: Payton Bradsky <pbradsky@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>