[far] Implement FAR validation in the C++ reader.
This changelist contains a port of the FAR validation functionality
from the Go and Rust readers to the C++ reader. It includes unit
testing for the dir name validation function and uses the same invalid
FARs to test the validation functionality as the Go and Rust readers.

Fixed: 55804
Change-Id: I6d22647136575a4794d6fa1026cb1c6cb8f1465f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461994
Reviewed-by: Ben Keller <galbanum@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>