[kernel][phys] Make "common_test" a GN build template.
The physboot build wants to generate a set of test targets using both
the "zbi_test" and the "phys_qemu_test" templates. Currently, this is
done using a loop that internally generates the appropriate targets.

As we have more variety in tests (such as tests that only work with an
input ZBI), the current approach is starting to become too inflexible.

Instead, create a new local GN template for a "common_test" that
generates both a "zbi_test" and "phys_qemu_test", longer term allowing
each test to have its own configuration, or to only build one of the two
environments.

Change-Id: Iadcaea5709d2e6e4795232ed57b99b76f86d8f08
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465097
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>