[fvm] Support the backup superblock in blobfs
fxrev.dev/459062 added a backup superblock to blobfs which is
only present in FVM images.

That CL made it so that newly created blobfs images would have this
backup superblock set up properly, but it did not do the same treatment
for the code in fvm/host which expands a blobfs image into an FVM image.
This meant that such images would not have a backup superblock
populated, which fails fsck. (Normal usage still works, since blobfs
does not consult the backup superblock in normal usage.)

Fixed this, and added two new tests:
- A test in fvm_sparse_image_reader_test which exercises loading a
  sparse FVM image and fscking its data/blobfs partitions
- A host test for fvm/host which performs a very basic smoke-test
  of generating a trivial image and fsck-ing said image

Change-Id: I73de71fa840685597671b69b2bfa7c53b877046a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467883
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>