[fxl] Fix invalid use of volatile in C++20
In C++20, volatile variables of non-class type cannot be used with
compound assignment operators (++, +=, |=, etc.) and also cannot be used
as the left-hand-side of an assignment where the result of the
assignment is used in a larger expression.

The latter type of usage appears in an FX_DCHECK in the destructor of
fxl::WeakPtrFactory, which has now been broken out into its own private
member function to satisfy the requirements of the C++20 standard.

TEST=no behavior change

Change-Id: I9cad4d338499278135eed4e30ebb3ab1c0cbe43b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/461725
Commit-Queue: Mark Schott <schottm@google.com>
Fuchsia-Auto-Submit: Mark Schott <schottm@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>