[MessageHub] Service MessageHub Definition.
This changelist introduces a new MessageHub intended to
replace all other MessageHub definitions. It is currently a
skeleton definition, to be populated as functionality is
migrated over to it.

Note that no tests have been added as the hub is not being
used. However, existing tests will exercise the hub once
changes land.

Fixed: 67490
Test: fx test -o setui_service_tests
Change-Id: Ib4bf93eccefa47f43575b4c19b4b7c40202a5b24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469612
Commit-Queue: Bryce Lee <brycelee@google.com>
Reviewed-by: William Xiao <wxyz@google.com>