[fint] Add gn_trace_path to set_artifacts.proto
This needs to be done as a separate commit from the commit that starts
setting the field – otherwise recipes will fail to deserialize the
textproto because the textproto format does not allow unknown fields.
We'll have to wait for this commit to be copied to the recipes repo
before starting to set the field in fint.

Bug: 65899
Change-Id: I8ab51e1c60c163a14e7a790e3121ea039bf18aed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470839
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>