[bt][hci] Fix race condition regression.
The previous CL introduced in a race condition in the test when fixing a memory leak.
The Event Arc must be consumed when it is used as a raw pointer in the
test to avoid it being double freed.

Test: Run bt-hci-lib-tests
Bug: fxbug.dev/67595
Change-Id: I01f0cd3b7b4d6f80a25efa087c4568b93f7d9204
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470821
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Jeff Belgum <belgum@google.com>
Fuchsia-Auto-Submit: Jeff Belgum <belgum@google.com>