[kernel][vm] Add inactive pager backed queue
Introduce a new page queue to track inactive userpager backed pages.
Pages are moved into this queue when they are explicitly promoted for
eviction, via the promote-no-clones command line for now, and with
eviction hints going forward (not yet implemented).

Pages are moved out of this queue when they are evicted, or when they
are accessed again which moves them to the first LRU page queue.
Inactive pages are considered for eviction before pages in the oldest
LRU queue.

This change only has effect if the kernel.page-scanner.promote-no-clones
command line is enabled. Instead of moving the promoted pages to the
tail of the oldest page queue, these pages will now be moved to the head
of the inactive queue, maintaining an LRU scheme across all inactive
pages.

Test: kernel-unittests

Bug: 65334
Change-Id: If25521a2bd671f357b6c3c07e7267d4f359eb1ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470430
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>