[flatland] Remove Validate Function
The Validate() function is no longer necessary
for the Renderer API and so it is now removed as
entirely, both as an API level function and as an
implementation detail for the subclassed renderers.
As a result, the BufferCollectionMetadata struct is
no longer needed as well and is likewise removed.

Testing: Existing tests are updated to account for
the API change.

Change-Id: Ied416bb4e3c5b3ffc96fe200946472c3e24b6e3d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466155
Commit-Queue: Chris Fontas <cfontas@google.com>
Reviewed-by: Evan Birenbaum <birenbaum@google.com>