[stdcompat]: Added traits.
Updated version and traits for properly selecting available features in
traits.
Removed 'callable_traits' until they are required by some library in
stdcompat. This seems to be here for fit::function and fit::promise
mostly.

Noted fit::function is similar to std::function it is a move only, and
not a replacement/polyfill for std::function.

Bug: 57349
Test: stdcompat-tests
Change-Id: I0cf119ef8d07d6170b5e7ae097cff351999362a2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/465008
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>