[stdcompat]: update internal/storage_internal
This just updats namespace and paths to match stdcompat(cpp17) related
style. The code is mostly as is. Unfortunately there are no strict tests
for this internal struct, but it's essentially a union type, used for
optional and variant implementation.

Though nothing is pulling this header right now, so further changes
might be necessary once optional is updated.

Bug: 57349
Change-Id: I167e7a0be7e276848403023a70c732c373336369
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466094
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>