[utest][core] Add concurrent vmo decommit test.
This CL adds a test that performs vmo decommit operations concurrently
with other threads writing to the same address.

It should not cause any problems to access a piece of memory that we are
decommitting.

Running this test with quemu in tcg-multithread mode, crashes the
kernel. See https://bugs.launchpad.net/qemu/+bug/1911351 for the qemu
bug report.

BUG=66978

Change-Id: I6c19691c4f6e3a21f000ed9d5d61cd93f8a017a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470740
Commit-Queue: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>