[httpsdate] Calculate poll offset from system UTC clock
Calculates, but does not yet report, the offset of the time polled from
the server and the system's UTC clock. This will serve as an
approximation of the error distribution for a single poll. Eventually
these metrics will be used to improve HTTPSDate's estimate of sample
error.

Bug: 63370

Change-Id: I814d634a0abef9bde40b5b73be7fc5c627ea82ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468678
Reviewed-by: Jody Sankey <jsankey@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>