[settings] Allow policy handlers to intercept and modify setting events
Policy handlers need to be able to intercept Changed events sent from
the setting handlers to the switchboard in order to adjust their values
according to policies.

When the policy logic was first designed, the Changed event caused the
recipient to send another Get request for the value. Now, the Changed
event contains the updated value, which eliminates a round trip, but
also was not intercepted by the policy layer.

Bug: 67148
Test: fx test setui_service_tests
Change-Id: Ie136b2d22fb5a2dad5593c7532d4e79b1defe399
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469214
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Bryce Lee <brycelee@google.com>