[stdcompat]: Rename in_place_internal -> utility.
This header provides a polyfill for constructor disambiguation tags for
optional and variant, std::in_place_*(t, index_t, etc).

This tags are part of the public API in FIT, just pulled from an
internal header transitively, there is no reason for this, since all the
content of this files are actually public.

The header will now be a polyfill for utility.

Bug: 57349
Test: stdcompat-tests
Change-Id: I2564d29141c08569ab42512a6248c1334dbe7ba4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/466098
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>