[debugger] Split out and enhance exception docs.
This splits out the existing two large documentation files into
topic-specific files. This will make it easier to modify and read.

The documentation hasn't really changed. Some parts of the documentation
should be enhanced and this is more obvious from this move. we can
enhance it in a separate pass.

Change-Id: I25f208e70c51089e00242252ebecb7089d8cd1e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467955
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>