[fostr] Support for printing uint8 and int8
This patch introduces support for printing uint8 and int8 members (for
structs, unions, tables, and vectors) in an ostream<< context.
Otherwise, these members are interpreted as a C-lang `char` type. Note
that `char` is not a FIDL primitive.

This patch also adds two small tests to verify arrays of primitives are
formatted correctly, such as array<float, 2> and array<int32, 2>. The
user can read the tests and thus look for the source of compilation
errors elsewhere (such as a missing `using` statement).

BUG: 67157

TEST=CQ

Change-Id: Ib7df745baffd475223aa95c03d66b90b4ebcc485
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469134
Commit-Queue: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>