[brcmfmac] Improve and count log throttling
With log throttling in place this change adds the ability to count the
number of times a log entry was throttled before it was logged again.
This is done by appending a string like '[Throttled 3 times]' if the
message was previously throttled, if no throttling occurred nothing is
appended.

This change also replaces the old throttling functionality that was
based on a counter to use the new throttling method.

BUG: 29355, 63984
Test: Included unit tests, manual testing with logs enabled/added

Change-Id: Ib42112b172bc26b9e7bccf3255419bef68afec67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470521
Commit-Queue: Bjoern Johansson <bjoernj@google.com>
Fuchsia-Auto-Submit: Bjoern Johansson <bjoernj@google.com>
Reviewed-by: Joshua Conner <joshconner@google.com>