[fint] Use JSON instead of textproto for set_artifacts
We use JSON instead of textproto for this message because it passes data
from fint back to the caller, but the source of truth for the proto
definition is in fint. Proto libraries don't have good support for
deserializing unknown fields in textprotos (unlike JSON protos), so if
we used textproto then we'd have to make two separate fint changes every
time we want to start setting a new field: one to add the field, then
another to set the field, which can only be landed after the updated
proto definition has been propagated to all consumers.

Serializing to binary rather than JSON would also work, but the
disadvantage of using binary is that we'd no longer be able to
arbitrarily delete and renumber fields of set_artifacts.proto, whereas
we *are* able to do that with static.proto and context.proto. This
discrepancy could get confusing – I expect it would be easy to forget
that it's not safe to arbitrarily modify set_artiacts.proto fields.

This is not safe to land until recipe change fxrev.dev/470803 lands.

Bug: 65899
Change-Id: I82bf47d583e09f0a7e22ff551cc8e73694c12874
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/470843
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Gary Boone <gboone@google.com>