[ffx][mdns] drop tasks for lost ip addresses
Linux does not error on sendmsg / recvmsg when the socket is bound to
an address no longer on the system, and worse even allow packets to
still be sent with the bindings source address (effectively spoofing
the source address). In order to avoid this outcome, explicily remove
tasks for sockets that are bound to no longer present addresses.

Bug: 67158
Change-Id: I724c65a81672abec281dfcdeb251ea4390f4f55d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/471719
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andrew Davies <awdavies@google.com>