[location] Send initial null region code
This change is the first of a couple changes to store and send the
initial regulatory region code on start up. If no region code is loaded
on boot, a null update would be sent.

This change adds the ability to send a null update through the API. Only
the first update sent can be null.

Test: fx test regulatory-region-tests
      fx test regulatory-region-integration-test
Bug: 66538
Change-Id: I09310472d1bc58eb3ce220e50242b94d6f273352
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/469604
Commit-Queue: Naomi Mccracken <nmccracken@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Kevin Sakuma <sakuma@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>