[swd] send errors to the syslog
Returning an error with `main() -> Result<(), Error>` emits logs to
STDOUT. Unfortunately STDOUT isn't sent to the syslog, so when these
processes are launched as a component the error is lost. This changes
`omaha-client`, `pkg-cache`, `pkg-resolver`, `system-update-checker` and
`system-update-committer` to also emit errors to the syslog.

Change-Id: I736c94451a09c486410fa51e2f273e3eceb4f6e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/468902
Reviewed-by: Kevin Wells <kevinwells@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>