[BUILD] Template for prebuilts with flavors (variants)
This is a template for selecting which flavor to use for packages that
are prebuilt into multiple flavorss that need to have the same runtime
component URLs.

It allows for the definition of multiple flavors for a prebuilt
package, and a build param which is used to easily allow a developer
to switch between those flavors.

As this can obscure which flavor is being used (the flavorss are no
longer distinguished in the build graph or component urls), the template
generates a manifest of what flavor has been chosen for each package,
so that build outputs are auditable.

This is meant to be transitional, while other, more comprehensive
tooling, is created to facilitate these sorts of needs.

Change-Id: Icf83210b654809a270b20bea9f56e8a9f74f5eff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/459507
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: Kevin Twidle <kpt@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>