[stdcompat] Port fit::string_view.
This implementation has a couple of buggy methods, that don't match the
specification. The issue is that the tests are incorrect. To simplify
review, this CL is merely getting this to compile in stdcompat.

The tests were moved to gtest, which has two effects:
 (*) Literal integer type must be specified. zxtest allows integer types
     to be promoted to a bigger one for comparisons.
 (*) EXPECT_STR_EQ -> EXPECT_STREQ.
 (*) EXPECT/ASSERT(args, message_format, message args) ->
     EXPECT_ASSERT(args) << message ags;

Bug: 57349
Test: stdcompat-tests
Change-Id: I71bf338b76a82f82837e2de221b59a4b5b2d8cfd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/467595
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>