[session] remove non-working input_session test and re-enable tests
As discussed in fxrev.dev/472069, the input_session test wasn't
compiling, and build targets had been ignoring it. The test, as
structured, was not relevant to the example binary, so it has been
removed, and replaced with the default "TBD" stand-in test.

Build targets are updated to re-enable the stand-in tests for
input_session and scene_manager_session examples.

Also renames the widely used noop test placeholder, and updates
references to `fx run-test` (deprecated in favor of `fx test`).

Fixed: 67160
Fixed: 67161
Change-Id: I7ce69f335b7971df761ed1ddff775ee456adbaf4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/472159
Commit-Queue: Rich Kadel <richkadel@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>